refactor Salaire.java and SalaireEntity.java

+++ studylink_api/src/main/java/com/devops/studylink/Salaire/repository/entities/SalaireEntity.java
@@ -27,5 +27,5 @@ public class SalaireEntity {
     private UUID id;
 
     @Column(name="name")
-    private Integer grossSalary;
+    private Double grossSalary;
 }
